refactor: move vault types to type folder
